+++ lib/src/scaffolds/src/utilities.dart
@@ -1,4 +1,4 @@
-part of scaffolds;
+part of '../scaffolds.dart';
 
 /// Shows a dialog to confirm the deletion of items (Can be single or multiple)
 /// Requires the following translations:
@@ -15,20 +15,31 @@ Future<bool> deleteConfirmationDialog({
   bool isCooldown = false,
   VoidCallback? onCooldown,
 }) async {
-  LayrzAppLocalizations i18n = LayrzAppLocalizations.of(context)!;
+  LayrzAppLocalizations? i18n = LayrzAppLocalizations.of(context);
+
+  String title = '';
+  String content = '';
+
+  if (isMultiple) {
+    title = i18n?.t('actions.confirmationMultiple.title') ?? 'Are you sure that you want to delete these items?';
+    content = i18n?.t('actions.confirmationMultiple.content') ?? 'Once deleted, you will not be able to recover them.';
+  } else {
+    title = i18n?.t('actions.confirmation.title') ?? 'Are you sure that you want to delete this item?';
+    content = i18n?.t('actions.confirmation.content') ?? 'Once deleted, you will not be able to recover it.';
+  }
 
   bool? result = await showDialog(
     context: context,
     builder: (context) {
       return ThemedDialog(
-        titleText: i18n.t('actions.${isMultiple ? 'confirmationMultiple' : 'confirmation'}.title'),
+        titleText: title,
         body: Text(
-          i18n.t('actions.${isMultiple ? 'confirmationMultiple' : 'confirmation'}.content'),
+          content,
           textAlign: TextAlign.center,
         ),
         dismiss: ThemedDialogAction(
           color: Colors.red,
-          labelText: i18n.t('actions.confirmation.dismiss'),
+          labelText: i18n?.t('actions.confirmation.dismiss') ?? 'Nevermind',
           isLoading: isLoading,
           isCooldown: isCooldown,
           onCooldown: onCooldown,
@@ -39,7 +50,7 @@ Future<bool> deleteConfirmationDialog({
           ThemedDialogAction(
             style: ThemedButtonStyle.filledTonal,
             color: Colors.green,
-            labelText: i18n.t('actions.confirmation.confirm'),
+            labelText: i18n?.t('actions.confirmation.confirm') ?? 'Do it!',
             isLoading: isLoading,
             isCooldown: isCooldown,
             onCooldown: onCooldown,