+++ lib/src/table/table.dart
@@ -1,16 +1,18 @@
-library scaffolds;
+library table;
+
+import 'dart:math';
 
 import 'package:flutter/material.dart';
 import 'package:layrz_models/layrz_models.dart';
 import 'package:layrz_theme/src/buttons/buttons.dart';
-import 'package:layrz_theme/src/dialogs/dialogs.dart';
 import 'package:layrz_theme/src/helpers/helpers.dart';
 import 'package:layrz_theme/src/inputs/inputs.dart';
-import 'package:layrz_theme/src/table/table.dart';
+import 'package:layrz_theme/src/scaffolds/scaffolds.dart';
 import 'package:layrz_theme/src/theme/theme.dart';
+import 'package:layrz_theme/src/tooltips/tooltips.dart';
 import 'package:material_design_icons_flutter/material_design_icons_flutter.dart';
 
-// Scaffolds
-part 'src/sidebar.dart';
-part 'src/cell.dart';
-part 'src/utilities.dart';
+part 'src/table.dart';
+part 'src/column.dart';
+part 'src/action.dart';
+part 'src/avatar.dart';